requirements.txt
gradio[mcp]>=4.0.0
pyalex>=0.13
PyYAML>=6.0
requests>=2.31.0
typing-extensions>=4.0.0
